% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_vapeplot_c}
\alias{scale_color_vapeplot_c}
\alias{scale_color_vapeplot_d}
\alias{scale_colour_vapeplot_c}
\alias{scale_colour_vapeplot_d}
\alias{scale_fill_vapeplot_c}
\alias{scale_fill_vapeplot_d}
\title{Vaporwave-aligned color maps}
\usage{
scale_color_vapeplot_c(option = "vaporwave", alpha = 1, reverse = FALSE,
  ...)

scale_color_vapeplot_d(option = "vaporwave", alpha = 1, reverse = TRUE,
  ...)

scale_colour_vapeplot_c(option = "vaporwave", alpha = 1, reverse = FALSE,
  ...)

scale_colour_vapeplot_d(option = "vaporwave", alpha = 1, reverse = TRUE,
  ...)

scale_fill_vapeplot_c(option = "vaporwave", alpha = 1, reverse = FALSE,
  ...)

scale_fill_vapeplot_d(option = "vaporwave", alpha = 1, reverse = TRUE,
  ...)
}
\arguments{
\item{option}{Choose from "vaporwave" (default), "cool", "crystal_pepsi",
"mallsoft", "jazzcup", "sunset", "macplus", or "seapunk".}

\item{alpha}{The alpha transparency, a number in [0,1]}

\item{reverse}{If \code{TRUE}, the direction of the colours is reversed.}

\item{...}{additional arguments to pass to \code{scale_color_gradientn} or
\code{discrete_scale}.}
}
\description{
The \code{vapeplot} scales provide continuous and discrete color schemes
found in the Python package
\href{https://github.com/dantaki/vapeplot}{vapeplot}.
}
\examples{
library(ggplot2)
library(vapeplot)

ggplot(diamonds) +
 geom_point(aes(x = carat, y = price, color = cut)) +
 scale_color_vapeplot_d("mallsoft")

ggplot(diamonds) +
  geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_vapeplot_d("seapunk")
}
