% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_palette}
\alias{show_palette}
\title{Display a color palette}
\usage{
show_palette(...)
}
\arguments{
\item{...}{the name of the palette or palettes; shows all palettes if no
argument is give.}
}
\description{
Display a palette in the graphics window by calling its name.
}
\examples{
library(vapeplot)
show_palette("cool", "sunset")
}
