% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{vapeplot}
\alias{vapeplot}
\alias{vapeplot_pal}
\title{Create a color ramp from a \code{vapeplot} palette}
\usage{
vapeplot(option = "vaporwave", n, alpha = 1, reverse = FALSE)

vapeplot_pal(option = "vaporwave", alpha = 1, reverse = FALSE)
}
\arguments{
\item{option}{Choose from "vaporwave" (default), "cool", "crystal_pepsi",
"mallsoft", "jazzcup", "sunset", "macplus", "seapunk".}

\item{n}{The number of colors (≥ 1) to be in the palette.}

\item{alpha}{The alpha transparency, a number in [0,1]}

\item{reverse}{If \code{TRUE}, the direction of the colours is reversed;
\code{FALSE} is the default.}
}
\description{
Create a color ramp from a \code{vapeplot} palette
}
\examples{
library(vapeplot)
image(volcano, col = vapeplot("sunset", 20))
}
