% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{vapeplot_palettes}
\alias{vapeplot_palettes}
\title{Vaporwave palettes}
\format{An object of class \code{list} of length 8.}
\usage{
vapeplot_palettes
}
\description{
Color palettes from the Python package
\href{https://github.com/dantaki/vapeplot}{vapeplot} - a matplotlib
extension.
}
\details{
Choose from several palettes:
\itemize{
\item vaporwave
\item cool
\item crystal_pepsi
\item mallsoft
\item jazzcup
\item sunset
\item macplus
\item seapunk
}
}
\examples{
library(vapeplot)
vapeplot_palettes$seapunk
}
\keyword{datasets}
